requirements.txt
Django==1.10.3
django-cors-headers==1.3.1
django-redis==4.6.0
djangorestframework==3.5.3
psycopg2==2.6.2
redis==2.10.5
Unidecode==0.4.19
virtualenv==15.1.0
